+++ Buffer.hpp
@@ -6,17 +6,17 @@ class Buffer{
 	uint64_t size;
 	unique_ptr<uint8_t> buffer;
 
-	uint32_t readPos, writePos;
+	uint64_t readPos, writePos;
 public:
 
 	Buffer(uint64_t size);
 	Buffer(const Buffer &buffer);
 	Buffer(Buffer &&buffer);
 
-	void write(const void *data, uint32_t length);
-	void write(istream &i, uint32_t length);
-	void read(void *dst, uint32_t length);
-	void read(ostream &o, uint32_t length);
+	void write(const void *data, uint64_t length);
+	void write(istream &i, uint64_t length);
+	void read(void *dst, uint64_t length);
+	void read(ostream &o, uint64_t length);
 
 
 	template<typename T>
@@ -34,8 +34,8 @@ public:
 
 	void dropPointers();
 
-	uint32_t getWriteSpace() const;
-	uint32_t getReadSpace() const;
+	uint64_t getWriteSpace() const;
+	uint64_t getReadSpace() const;
 
 	void *get();
 	const void *get() const;
